refactor: replace normalize_height by distance_to_dtm
