+++ src/pointtree/operations/__init__.py
@@ -3,7 +3,7 @@
 from ._cloth_simulation_filtering import *
 from ._create_digital_terrain_model import *
 from ._estimate_with_linear_model import *
-from ._normalize_height import *
+from ._distance_to_dtm import *
 from ._points_in_ellipse import *
 from ._polygon_area import *
